requirements.txt
aiogram==2.12.1
aiohttp==3.7.4.post0
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
Babel==2.9.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-adminlte2==0.4.3
django-mdeditor==0.1.18
django-solo==1.1.5
environs==8.0.0
hiredis==2.0.0
idna==2.10
loader==2017.9.11
Markdown==3.3.4
marshmallow==3.10.0
multidict==5.1.0
Pillow==8.1.2
psycopg2==2.8.6
pyTelegramBotAPI==3.7.6
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
soupsieve==2.2
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
